+++ lib/src/assets_config/local_image_asset.dart
@@ -1,6 +1,6 @@
-import 'package:fl_ui_config/src/widget/stateful_widget_with_ui_config.dart' show brightness, isHighContrastEnabled;
-import 'package:flutter/widgets.dart' show BoxFit, BuildContext, Brightness, Image, Widget;
-import 'package:flutter_svg/flutter_svg.dart' show SvgPicture;
+import 'package:fl_ui_config/src/widget/stateful_widget_with_ui_config.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/flutter_svg.dart';
 
 import 'asset.dart';